pyproject.toml
[tool.poetry]
name = "Pengenuity"
version = "0.1.0"
description = ""
authors = ["Dory <info@penguins-lab.com>"]

[tool.poetry.dependencies]
python = ">=3.10.10,<3.12"
openai = "^0.27.0"
langchain = "^0.0.144"
python-dotenv = "^1.0.0"
google-api-python-client = "^2.84.0"
faiss-cpu = "^1.7.3"
sentence-transformers = "^2.2.2"
jsonschema = "^4.17.3"
marvin = "^0.7.5"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"